fix: fix config form prevent focus from working
