+++ src/components/NoData.tsx
@@ -1,4 +1,4 @@
-import { IconWithText, XIcon } from 'components/FeatherIcon'
+import { IconWithText, XIcon } from 'components/FeatherIcon.js'
 
 export const NoData = ({ children }: { children?: string }) => (
 	<div
